requirements.txt
polymarket-apis>=0.5.0
python-dotenv>=1.0.0
rich>=13.0.0
